+++ config/default.yaml
@@ -48,7 +48,7 @@ services:
   draftStore:
     redis:
       host: 'localhost'
-      port: 6379
+      port: 6380
       key:
       tls: false
 health:
