--- draft-store.yml
@@ -1,12 +1,0 @@
-version: '2'
-services:
-  draft-store:
-    image: redis:6.2.6-alpine
-    ports:
-      - '6379:6379'
-    command: redis-server --save 20 1 --loglevel warning
-    volumes:
-      - cache:/data
-volumes:
-  cache:
-    driver: local
