+++ src/main/modules/draft-store/index.ts
@@ -24,10 +24,10 @@ export class DraftStoreClient {
     const client = new Redis({
       host: config.get('services.draftStore.redis.host'),
       port: config.get('services.draftStore.redis.port'),
-      // password: config.get('session.redis.key'),
-      // tls: {
-      //   servername: config.get('services.draftStore.redis.host'),
-      // },
+      password: config.get('session.redis.key'),
+      tls: {
+        servername: config.get('services.draftStore.redis.host'),
+      },
     });
     Logger.error('connection string');
