+++ src/main/steps/applicant1/address/controller/ConfirmChangePostController.ts
@@ -1,0 +1,33 @@
+import autobind from 'autobind-decorator';
+import { Response } from 'express';
+
+import { FieldPrefix } from '../../../../app/case/case';
+import { AppRequest } from '../../../../app/controller/AppRequest';
+import { AnyObject, PostController } from '../../../../app/controller/PostController';
+import { Form, FormFields, FormFieldsFn } from '../../../../app/form/Form';
+
+@autobind
+export default class ConfirmChangeAddressController extends PostController<AnyObject> {
+  constructor(protected readonly fields: FormFields | FormFieldsFn, protected readonly fieldPrefix: FieldPrefix) {
+    super(fields);
+  }
+
+  public async post(req: AppRequest<AnyObject>, res: Response): Promise<void> {
+    const fields = typeof this.fields === 'function' ? this.fields(req.session.userCase) : this.fields;
+    const form = new Form(fields);
+    const { saveAndSignOut, saveBeforeSessionTimeout, _csrf, ...formData } = form.getParsedBody(req.body);
+
+    req.session.errors = form.getErrors(formData);
+
+    req.session.returnUrl = undefined;
+    req.session.userCase.checkYourAnswersReturn = false;
+    req.session.userCase.changeAddressBothApplicants = undefined;
+
+    if (req.session.errors.length > 0) {
+      this.redirect(req, res);
+      return;
+    }
+
+    this.redirect(req, res);
+  }
+}
